Generate Fibonacci sequence up to a given number of terms
This program generates and displays the Fibonacci sequence up to the number of terms specified by the user. It initializes the first two terms of the sequence, then calculates and prints the next terms in the sequence iteratively. The sequence is displayed with each term separated by a comma. The user is prompted to input the number of terms to display.