Add marriage status checker using user input code
This program continuously prompts the user to enter a marriage status code until the code '1' is entered. It uses a switch-case structure to evaluate the input. If the user enters 'm' or 'M', it outputs "Individual is married." For 'd' or 'D', it outputs "Individual is divorced," and for 'w' or 'W', it outputs "Individual is widowed." If an invalid code is entered, the program will output "Invalid code." The loop continues until the user enters '1', which terminates the program.