Generate a number-star pattern based on user input
This program generates a pattern based on a user-input integer n between 1 and 9. For each row i, the program prints numbers from 1 to i, followed by stars (*) until the total number of characters (numbers and stars) in the row equals n. The pattern decreases the number of stars in each subsequent row while increasing the number of printed numbers. The user is prompted to input an integer, and the program then prints the corresponding pattern based on the input value.