Calculate and assign grades based on Z-scores
This program calculates the mean, standard deviation, and Z-scores for a set of student marks, and then assigns letter grades based on the Z-scores. The user inputs the total number of students and their respective marks. The program calculates the mean and standard deviation of the marks, then computes the Z-score for each student. Based on the Z-scores, it assigns a grade as follows:

A for Z-score ≥ 1.50
B for Z-score between 1.00 and 1.49
C for Z-score between 0.00 and 0.99
D for Z-score between -1.00 and -0.99
E for Z-score < -1.00
The program also outputs the highest mark and the calculated mean and standard deviation of the marks. Results are displayed with two decimal places of precision.