Added C++ program with multiple functions for personalized greeting messages
This C++ program demonstrates the use of functions to print personalized greeting messages. It includes the following:

printmsg(): Prints a generic "Welcome!!!" message.
printmsg1(string n): Prints a personalized greeting, such as "Good morning!!! [name]", where the name is passed as an argument.
printmsg2(): Prints a generic "Good afternoon!!!" message.
The program currently calls the printmsg1() function with a sample name ("Ram") to demonstrate its functionality. The other functions are defined but not called in the main program.

This code can be used as an introduction to function usage and parameter passing in C++.